User Microservice to List Profile Updates

+++ src/main/java/com/mtech/sjmsuser/entity/UserProfile.java
@@ -9,8 +9,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.List;
-
 @Getter
 @Setter
 @Entity
@@ -20,7 +18,7 @@ public class UserProfile {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer Id;
+    private Long Id;
     private String email;
     private String hashedPassword;
     private String role;
